+++ src/main/java/com/liteisle/common/domain/response/ShareRecordPageResp.java
@@ -8,8 +8,8 @@ import java.util.List;
 @Data
 public class ShareRecordPageResp {
     private Long total;
-    private Integer currentPage;
-    private Integer pageSize;
+    private Long currentPage;
+    private Long pageSize;
     private List<ShareRecord> records;
     
     @Data
